updates and removes utils class for specific functionality
