+++ lib/app.rb
@@ -3,7 +3,11 @@ require 'sinatra'
 require 'sinatra/json'
 require 'rack/parser'
 require 'pry'
+
+require 'polypeptide'
 require 'protein'
 require 'protein_api'
 require 'protein_utils'
+require 'response'
+require 'validator'
