+++ lib/response.rb
@@ -1,0 +1,10 @@
+class Response
+  def self.not_correct_chars
+    'Input must be a codon.'
+  end
+
+  def self.not_valid
+    'Not a valid RNA sequence.'
+  end
+
+end
